A tweak for MSVC6 : If gmu_DO_PASS_FULLPATH_TO_COMPILER=1, the .c/.cpp path passed to cl.exe will get all "/" converted to "\" . For example:
  ../src/foo.cpp

converted to 

  D:\myprj\src\foo.cpp

Without this tweak, "../src/foo.cpp" will be recorded into PDB literally, and VS2008(VS2005, VS2003 as well maybe) will fail to recognize the "foo.cpp" filename, then, your .c/cpp source files will not be found when debugging.

VS2003 and above do not need this tweak, and are not affected by this commit.

git-svn-id: file:///mnt/i/1027-HOME-svnreps/nlscan-svnreps/makingsys/GnumakeUniproc/trunk@1160 429491bc-2e6e-d343-9161-6a1f6e34ed84